requirements.txt
asgiref==3.5.2
azure-core==1.26.1
azure-storage-blob==12.14.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==3.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pycparser==2.21
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
urllib3==1.26.12
